cmake: warnings as errors
The library fails to build when `-werror` is enabled.
The errors are rather simple to fix while utilizing explicit casts from
json.

This commit enables warnings as errors and fixes the errors cause by
this change.

motivation: modern codebases should threat warnings as errors.
One solution is to disable locally for a given library.
Another is to fix these warnings in the library itself.

+++ src/operations/data/var.cpp
@@ -44,7 +44,7 @@ namespace json_logic
                     if (!current.is_array())
                         throw JsonLogicException(__FUNCTION__, "Data object is not an array, but index was provided");
 
-                    if (index >= current.size())
+                    if (static_cast<std::size_t>(index) >= current.size())
                         throw JsonLogicException(__FUNCTION__, "Array index out of bounds");
 
                     current = current[index];
